db_set and db_set_function headers
