+++ cs/CMakeLists.txt
@@ -70,6 +70,8 @@ set (COMPAT_HEADERS
   ${COMPAT_DIR}/db_json_allocator.hpp
   ${COMPAT_DIR}/db_json_path.hpp
   ${COMPAT_DIR}/db_json_types_internal.hpp
+  ${COMPAT_DIR}/db_set.h
+  ${COMPAT_DIR}/db_set_function.h
   ${COMPAT_DIR}/dbtype_function.h
   ${COMPAT_DIR}/dbtype_function.i
   ${COMPAT_DIR}/db_admin.h
