+++ src/base/object_representation.h
@@ -39,6 +39,7 @@
 #endif /* !WINDOWS */
 
 #include "byte_order.h"
+#include "db_set.h"
 #include "error_manager.h"
 #include "memory_alloc.h"
 #include "oid.h"
@@ -1117,22 +1118,7 @@ struct db_reference
 typedef struct setobj SETOBJ;
 
 typedef struct db_set SETREF;
-struct db_set
-{
-  /*
-   * a garbage collector ticket is not required for the "owner" field as
-   * the entire set references area is registered for scanning in area_grow.
-   */
-  struct db_object *owner;
-  struct db_set *ref_link;
-  struct setobj *set;
-  char *disk_set;
-  DB_DOMAIN *disk_domain;
-  int attribute;
-  int ref_count;
-  int disk_size;
-  need_clear_type need_clear;
-};
+
 
 #if defined (__cplusplus)
 class JSON_VALIDATOR;
